lib: move Rayon wrappers to lib crate
