+++ common/cf-solana/src/lib.rs
@@ -4,6 +4,6 @@ extern crate alloc;
 #[cfg(any(feature = "std", test))]
 extern crate std;
 
+mod blake3;
 pub mod proof;
 pub mod types;
-mod utils;
