+++ common/lib/Cargo.toml
@@ -6,10 +6,11 @@ edition = "2021"
 
 [dependencies]
 base64.workspace = true
-bs58 = { workspace = true, optional = true }
 borsh = { workspace = true, optional = true }
+bs58 = { workspace = true, optional = true }
 bytemuck = { workspace = true, features = ["derive"] }
 derive_more.workspace = true
+rayon = { workspace = true, optional = true }
 sha2.workspace = true
 solana-program = { workspace = true, optional = true }
